LayoutTests/platform/ios-simulator/webaudio/delaynode-expected.txt

Tests attribute and basic functionality of Delay.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS delay.numberOfInputs === 1 is true
PASS delay.numberOfOutputs === 1 is true
PASS delay.delayTime.defaultValue === 0.0 is true
PASS delay.delayTime.minValue === 0.0 is true
PASS delay.delayTime.maxValue === 1.0 is true
PASS delay.delayTime.value === 0.0 is true
PASS delay.delayTime.value === 0.5 is true
PASS Test signal was correctly delayed by 0.5 sec.
PASS successfullyParsed is true

TEST COMPLETE
